% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisonpicsuite-package.R
\docType{package}
\name{bisonpicsuite-package}
\alias{bisonpicsuite}
\alias{bisonpicsuite-package}
\title{bisonpicsuite: Easily Install and Load `bisonpic` R Packages}
\description{
This suite of packages is designed to facilitate data exploration and analysis of wood bison camera trap data, using underlying functions to visualize and model data (`bisonpictools`) and Shiny applications that streamline the data visualization (`shinybisonpic`) and data analysis (`runbisonpic`) processes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/bisonpicsuite}
  \item \url{https://poissonconsulting.github.io/bisonpicsuite/}
  \item Report bugs at \url{https://github.com/poissonconsulting/bisonpicsuite/issues}
}

}
\author{
\strong{Maintainer}: Nicole Hill \email{nicole@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7623-2153}{ORCID})

Authors:
\itemize{
  \item Ayla Pearson \email{ayla@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-7388-1222}{ORCID})
  \item Andrea Kortello \email{andrea@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-8047-3331}{ORCID})
}

Other contributors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID}) [contributor]
  \item Troy Hegel \email{Troy.Hegel@gov.ab.ca} [contributor]
  \item Province of Alberta [copyright holder]
}

}
\keyword{internal}
