% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bisonpicsuite_packages}
\alias{bisonpicsuite_packages}
\title{List all bisonpicsuite packages}
\usage{
bisonpicsuite_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include bisonpicsuite in the list?}
}
\description{
Inspired by \code{tidyverse::tidyverse_packages()}.
}
\examples{
bisonpicsuite_packages()
}
